# src/assets/open-api/food-catalog-swagger-v2.yaml
swagger: "2.0"
host: food-catalog-api-dev.azurewebsites.net
schemes:
  - https
info:
  title: Food-Inventory
  version: v1
  description: An Api that allows to manage food stock for a restaurant
definitions:
  FoodDTO:
    type: object
    properties:
      name:
        type: string
        x-nullable: true
      price:
        format: double
        type: number
      inStock:
        format: int32
        type: integer
      minStock:
        format: int32
        type: integer
      pictureUrl:
        type: string
        x-nullable: true
      description:
        type: string
        x-nullable: true
    additionalProperties: false
  FoodItem:
    type: object
    properties:
      id:
        format: int32
        type: integer
      name:
        type: string
        x-nullable: true
      price:
        format: double
        type: number
      inStock:
        format: int32
        type: integer
      minStock:
        format: int32
        type: integer
        description: Minimum amount to keep in stock
      pictureUrl:
        type: string
        x-nullable: true
      description:
        type: string
        x-nullable: true
    additionalProperties: false
x-ms-openai-manifest:
  humanName: Food-Inventory
  modelDescription: An Api that allows to manage food stock for a restaurant
  contactEmail: ""
  legalUrl: ""
securityDefinitions: {}
paths:
  /Config: {}
  /Config/getAllEnv: {}
  /Food:
    get:
      tags:
        - Food
      produces:
        - text/plain
        - application/json
        - text/json
      responses:
        "200":
          description: OK
          schema:
            type: array
            items:
              $ref: "#/definitions/FoodItem"
      x-ms-openai-data:
        openai-enabled: true
        operations:
          - x-ms-require-user-confirmation: false
      summary: Gets all food
      description: Returns every food item currently available in the catalog.
      operationId: GetFood
    post:
      tags:
        - Food
      consumes:
        - application/json
        - text/json
        - application/*+json
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - in: body
          name: body
          schema:
            $ref: "#/definitions/FoodDTO"
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/FoodItem"
      x-ms-openai-data:
        openai-enabled: true
        operations:
          - x-ms-require-user-confirmation: false
      summary: Add food
      description: Creates a new food item in the catalog.
      operationId: AddFood
    put:
      tags:
        - Food
      consumes:
        - application/json
        - text/json
        - application/*+json
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - in: body
          name: body
          schema:
            $ref: "#/definitions/FoodItem"
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/FoodItem"
      x-ms-openai-data:
        openai-enabled: true
        operations:
          - x-ms-require-user-confirmation: false
      summary: UpdateFood
      description: Updates an existing food item with the provided details.
      operationId: UpdateFood
  /Food/byname:
    get:
      tags:
        - Food
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - in: query
          name: name
          type: string
      responses:
        "200":
          description: OK
          schema:
            type: array
            items:
              $ref: "#/definitions/FoodItem"
      x-ms-openai-data:
        openai-enabled: true
        operations:
          - x-ms-require-user-confirmation: false
      summary: GetFoodByName
      description: Finds food items whose name matches the supplied query.
      operationId: GetFoodByName
  /Food/{id}:
    get:
      tags:
        - Food
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - in: path
          name: id
          required: true
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/FoodItem"
      x-ms-openai-data:
        openai-enabled: true
        operations:
          - x-ms-require-user-confirmation: false
      summary: GetFoodByID
      description: Fetches a single food item by its unique identifier.
      operationId: GetFoodByID
    delete:
      tags:
        - Food
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - in: path
          name: id
          required: true
          type: integer
          format: int32
      responses:
        "200":
          description: OK
      x-ms-openai-data:
        openai-enabled: true
        operations:
          - x-ms-require-user-confirmation: false
      summary: DeleteFood
      description: Removes a food item from the catalog.
      operationId: DeleteFood
  /Food/{id}/update-stock:
    patch:
      tags:
        - Food
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - in: path
          name: id
          required: true
          type: integer
          format: int32
        - in: query
          name: amount
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/FoodItem"
      x-ms-openai-data:
        openai-enabled: true
        operations:
          - x-ms-require-user-confirmation: false
      summary: UpdateFoodStock
      description: Adjusts the in-stock quantity for a specific food item.
      operationId: UpdateFoodStock