# src/food-catalog-api/Controllers/FoodController.cs
using System.Collections.Generic;
using System.Linq;
using Microsoft.AspNetCore.Mvc;
using Microsoft.EntityFrameworkCore;
using Microsoft.Extensions.Configuration;
using FoodApp;

namespace FoodApi
{
    [Route("[controller]")]
    [ApiController]
    public class FoodController : ControllerBase
    {
        public FoodController(FoodDBContext context, IConfiguration config)
        {
            ctx = context;
            cfg = config.Get<FoodConfig>();
        }

        FoodDBContext ctx;
        FoodConfig cfg;

        [HttpGet()]
        public IEnumerable<FoodItem> GetFood()
        {
            return ctx.Food.ToArray();
        }

        [HttpGet("byname")]
        public ActionResult<IEnumerable<FoodItem>> GetFoodByName([FromQuery] string name)
        {
            if (string.IsNullOrWhiteSpace(name))
                return BadRequest("Name parameter is required.");
            var items = ctx.Food.Where(f => f.Name.Contains(name)).ToList();
            if (items.Count == 0)
                return NotFound();
            return Ok(items);
        }

        [HttpGet("{id}")]
        public FoodItem GetById(int id)
        {
            return ctx.Food.FirstOrDefault(v => v.ID == id);
        }

        [HttpPost()]
        public FoodItem InsertFood(FoodDTO item)
        {
            var foodItem = new FoodItem
            {
                Name = item.Name,
                Price = item.Price,
                InStock = item.InStock,
                PictureUrl = item.PictureUrl,
                Description = item.Description
            };
            ctx.Food.Add(foodItem);
            ctx.SaveChanges();
            return foodItem;
        }

        [HttpPut()]
        public FoodItem UpdateFood(FoodItem item)
        {
            ctx.Food.Attach(item);
            ctx.Entry(item).State = EntityState.Modified;
            ctx.SaveChanges();
            return item;
        }

        [HttpDelete("{id}")]
        public ActionResult Delete(int id)
        {
            var item = GetById(id);
            if (item != null)
            {
                ctx.Remove(item);
                ctx.SaveChanges();
            }
            return Ok();
        }

        [HttpPatch("{id}/update-stock")]
        public ActionResult<FoodItem> UpdateInStock(int id, [FromQuery] int amount)
        {
            var item = ctx.Food.FirstOrDefault(f => f.ID == id);
            if (item == null)
                return NotFound();
            item.InStock += amount;
            ctx.SaveChanges();
            return Ok(item);
        }
    }
}